refactor: begin ADR-061 Phase 3 by moving artifact_plan.py into workflows
Move the flat abicheck/artifact_plan.py (ResolvedArtifactPlan, the shared
cleanup-thunk session type dump/scan already use) into
abicheck.workflows.artifact.contracts, the contracts.py half of Phase 3's
target ArtifactRequest -> ResolvedArtifactPlan -> ArtifactResult layout.
The module had zero first-party imports, so this is a pure location move:
its four call sites (service_dump_pipeline.py, service_input_resolution.py,
cli_dump_helpers.py, cli_dump_non_elf.py) now import it from the new
package and are otherwise unchanged.

Document the next blocker in ADR-061 itself: those four call sites' owning
service modules cannot yet move into workflows/ because they reach into
cli_dump_helpers.py (a frontends-destined module), which would invert the
workflows -> frontends dependency direction the same way Phase 2's
compare -> policy coupling did.
