+++ abicheck/workflows/artifact/__init__.py
@@ -1,0 +1,7 @@
+"""Canonical implementation owner for the shared artifact-resolution contract."""
+
+from .contracts import ResolvedArtifactPlan
+
+__all__ = [
+    "ResolvedArtifactPlan",
+]
