+++ tests/test_artifact_plan.py
@@ -12,8 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for :mod:`abicheck.artifact_plan` — Phase 1 (Milestones A and B) of
-the duplication-and-convergence plan's ``ResolvedArtifactPlan`` primitive."""
+"""Tests for :mod:`abicheck.workflows.artifact.contracts` — Phase 1
+(Milestones A and B) of the duplication-and-convergence plan's
+``ResolvedArtifactPlan`` primitive."""
 
 from __future__ import annotations
 
@@ -21,7 +22,7 @@ from pathlib import Path
 
 import pytest
 
-from abicheck.artifact_plan import ResolvedArtifactPlan
+from abicheck.workflows.artifact import ResolvedArtifactPlan
 
 
 def test_normal_exit_runs_cleanups_once():
@@ -274,6 +275,6 @@ def test_resolve_dump_request_attaches_a_matching_artifact_plan(
     assert plan.public_headers == resolved.public_headers
     assert plan.public_header_dirs == resolved.public_header_dirs
     # Resolution allocates no resource today (the L3->L2 fold stays inside
-    # execute -- see artifact_plan.py's own module docstring), so the
-    # attached plan must start with nothing to clean up.
+    # execute -- see workflows/artifact/contracts.py's own module
+    # docstring), so the attached plan must start with nothing to clean up.
     assert plan.pending_cleanups == []
